fix(dbus): clear the eight warnings in the D-Bus subsystem
Invisible on an incremental build: the objects were already up to date,
so nothing reported them until cmacs/dbus/*.c was touched.

The mechanical four.  Two casts to int of a pid_t that IS int, widened
to long with a matching %ld instead -- a format that only happens to
match the platform's pid_t is a quiet breakage waiting for a platform
where it does not.  Two syms_of_* with no prototype: the externs lived
in cmacs-dbus.c, where the translation units that DEFINE them could not
see them, and now live in cmacs-dbus-internal.h, which both include.

The four set-but-unused needed reading:

  * cintrospect List built a three-element args[] and then passed a
    five-element list inline.  The array was stale, not merely unused --
    an empty glob has to reach elisp as a bare nil rather than an empty
    string, so the quotes around it are themselves substituted, and the
    call grew two entries the array never got.  Deleted; the conditional
    list belongs at the call.

  * clipboard Put built args[1] and passed { text, text } inline.  Here
    the array was simply one short: the template names TEXT twice, once
    for the kill-ring and once for the selection.  Widened and used, so
    the branch reads like every other one in the file.

  * monitor and compositor each declared a GError *err in on_method that
    every caller appears to pass as &err -- but RETURN_STR substitutes
    its argument INSIDE a block that declares an err of its own, so that
    is what &err has always bound to, and what the reply reads back.  No
    error was being lost; the function-level one was shadowed and dead.
    Deleted, with a note on the macro so it does not come back.

Verified with 'touch cmacs/dbus/*.c && make -j8': no compiler warnings
outside upstream's src/xwidget.c.  make -C test check-cmacs stays at
2281 as expected, 0 unexpected.

+++ cmacs/dbus/cmacs-dbus-iface-instance.c
@@ -95,7 +95,7 @@ on_method_call (GDBusConnection *c, const gchar *s, const gchar *o,
       uptime = cmacs_dispatch_eval_string ("(emacs-uptime)", NULL);
 
       out = g_string_new ("{");
-      g_string_append_printf (out, "\"pid\":%d", (int) getpid ());
+      g_string_append_printf (out, "\"pid\":%ld", (long) getpid ());
 
       g_string_append (out, ",\"version\":");
       json_append_string (out, version != NULL ? version : PACKAGE_VERSION);
